requirements.txt
# Core dependencies
aiohttp>=3.8.5
asyncio>=3.4.3
fastapi>=0.95.0
loguru>=0.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.4.2
uvicorn>=0.22.0
websockets>=11.0.0

# LLM providers
openai>=1.0.0
anthropic>=0.5.0

# Memory
faiss-cpu>=1.7.4
pgvector>=0.2.0
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.0

# Tools
beautifulsoup4>=4.12.2
duckduckgo-search>=3.8.0
playwright>=1.38.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0
